Feature/optimize division (#701)
* Replaced as many divisions as possible with floting operations, didn't check value range, yields ~5-6 raw Synth voices

* dbt format

* Converted all floats to double

* Format

+++ src/deluge/processing/audio_output.cpp
@@ -220,7 +220,7 @@ renderEnvelope:
 			inputChannelNow = AudioInputChannel::NONE; // 0 means combine channels
 		}
 
-		int32_t amplitudeIncrement = (amplitudeAtEnd - amplitudeAtStart) / numSamples;
+		int32_t amplitudeIncrement = (int32_t)((double)(amplitudeAtEnd - amplitudeAtStart) / (double)numSamples);
 		int32_t amplitudeNow = amplitudeAtStart;
 
 		do {
